chore: cleanup- moved serialise prediction to services
