+++ ml_classifier/classifier/views.py
@@ -4,21 +4,13 @@ from django.http import JsonResponse, HttpResponseBadRequest, HttpResponseNotAll
 from django.shortcuts import render
 from django.views.decorators.csrf import csrf_exempt
 from django.contrib.auth.decorators import login_required, permission_required
-from ml.predict import predict
-from classifier.models import Prediction
+from ml_classifier.ml.predict import predict
+from models import Prediction
+from services import read_logs, serialise_prediction
 from PIL import Image
 from io import BytesIO
+
 # Create your views here.
-
-
-def serialise_prediction(prediction: Prediction) -> dict:
-    return {
-        "predicted_at": prediction.predicted_at,
-        "user": prediction.user,
-        "imagenet_class": prediction.imagenet_class,
-        "probability": prediction.probability
-    }
-
 
 
 @login_required
